refactor: Follow codebase standard for Dialogs
