+++ src/layout/Footer/FooterHost.svelte
@@ -10,38 +10,14 @@
 	import { RV_DMCA_GUID } from '$env/static/public';
 	import { onMount } from 'svelte';
 	import Divider from '$lib/components/Divider.svelte';
-	import { email } from '$data/api/settings';
-	import Dialog from '$layout/Dialogs/Dialog.svelte';
 	import Button from '$lib/components/Button.svelte';
-	import Input from '$lib/components/Input.svelte';
+	import EmailDialog from '$layout/Dialogs/EmailDialog.svelte';
 
 	const aboutQuery = createQuery(queries.about());
 
 	let location: string;
 
 	let showEmailDialog = false;
-	let enableInputSeconds = 15;
-	let keyword = '';
-
-	$: interval = 0;
-
-	function openDialog() {
-		showEmailDialog = true;
-
-		if (enableInputSeconds > 0) {
-			interval = setInterval(() => {
-				enableInputSeconds -= 1;
-				if (enableInputSeconds <= 0) clearInterval(interval);
-			}, 1000);
-		}
-	}
-
-	function closeDialog() {
-		showEmailDialog = false;
-		clearInterval(interval);
-		enableInputSeconds = 15;
-		keyword = '';
-	}
 
 	onMount(() => {
 		// DMCA Protection Badge
@@ -49,32 +25,7 @@
 	});
 </script>
 
-<Dialog bind:dialogOpen={showEmailDialog} notDismissible>
-	<svelte:fragment slot="title">Abuse notice</svelte:fragment>
-	<svelte:fragment slot="description">
-		<p>
-			This E-Mail address is <b>not</b> for support, help, bug reports or feature requests. It must
-			have a subject and body and have the keyword <span style="user-select: none;">'Reficio'</span>
-			in either, otherwise your mail will be
-			<b>ignored</b>.
-		</p>
-		<br />
-		{#if enableInputSeconds == 0}
-			<span>Enter the keyword, then click "Send".</span>
-			<br />
-			<br />
-			<Input placeholder="Keyword" type="text" bind:value={keyword}></Input>
-		{:else}
-			<span>Read above and wait <b>{enableInputSeconds}</b> seconds.</span>
-		{/if}
-	</svelte:fragment>
-	<svelte:fragment slot="buttons">
-		{#if keyword.toLowerCase() === 'reficio'}
-			<a href="mailto:{email()}">Send</a>
-		{/if}
-		<Button type="filled" on:click={closeDialog}>Close</Button>
-	</svelte:fragment>
-</Dialog>
+<EmailDialog bind:showEmailDialog />
 
 <Divider horizontalPadding={'15px'} />
 <footer in:fly={{ y: 10, easing: quintOut, duration: 750 }}>
@@ -115,8 +66,10 @@
 	<div class="bottom">
 		<div id="logo-name"><span>Re</span>Vanced</div>
 		<a href="/donate"><div>Donate</div></a>
-		<Button type="text" style="color: var(--text-four); font-weight: 600;" on:click={openDialog}
-			>E-Mail</Button
+		<Button
+			type="text"
+			style="color: var(--text-four); font-weight: 600;"
+			on:click={() => (showEmailDialog = true)}>E-Mail</Button
 		>
 		<!-- DMCA Protection Badge -->
 		<a
